tests/wpt/web-platform-tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_flex-shorthand-initial.html

<!DOCTYPE html>
<title>flexbox | computed style | flex: initial</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#flex-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	flex: initial;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-grow"),
	          "0");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-shrink"),
	          "1");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-basis"),
	          "auto");
});
</script>
